rust: move loader diff to new module
Also fix build errors
