+++ ffi/rust/firedancer-diff/Cargo.toml
@@ -14,9 +14,8 @@ fuzz-asan = ["firedancer-sys/fuzz-asan"]
 diff = "0.1.13"
 hxdmp = "0.2.1"
 firedancer-sys = { path = "../firedancer-sys", version = "0.3.0" }
-solana_rbpf = "0.5.0"
-solana-bpf-loader-program = { git = "https://github.com/solana-labs/solana", rev = "a674ffc9bda0c4535812a1c5e37398db9ba7be53", version = "1.16.0" }
-solana-logger = { git = "https://github.com/solana-labs/solana", rev = "a674ffc9bda0c4535812a1c5e37398db9ba7be53", version = "1.16.0" }
-solana-metrics = { git = "https://github.com/solana-labs/solana", rev = "a674ffc9bda0c4535812a1c5e37398db9ba7be53", version = "1.16.0" }
-solana-program-runtime = { git = "https://github.com/solana-labs/solana", rev = "a674ffc9bda0c4535812a1c5e37398db9ba7be53", version = "1.16.0" }
-solana-sdk = { git = "https://github.com/solana-labs/solana", rev = "a674ffc9bda0c4535812a1c5e37398db9ba7be53", version = "1.16.0" }
+solana-bpf-loader-program = { git = "https://github.com/solana-labs/solana", rev = "91186d38600f2c5acd3f1126aaedc2348368397a", version = "1.17.0" }
+solana-logger = { git = "https://github.com/solana-labs/solana", rev = "91186d38600f2c5acd3f1126aaedc2348368397a", version = "1.17.0" }
+solana-metrics = { git = "https://github.com/solana-labs/solana", rev = "91186d38600f2c5acd3f1126aaedc2348368397a", version = "1.17.0" }
+solana-program-runtime = { git = "https://github.com/solana-labs/solana", rev = "91186d38600f2c5acd3f1126aaedc2348368397a", version = "1.17.0" }
+solana-sdk = { git = "https://github.com/solana-labs/solana", rev = "91186d38600f2c5acd3f1126aaedc2348368397a", version = "1.17.0" }
