+++ ffi/rust/firedancer-sys/Cargo.toml
@@ -24,4 +24,4 @@ fuzz-asan = []
 paste = "1.0.12"
 
 [build-dependencies]
-bindgen = { version = "0.66.0", features = ["experimental"] }
+bindgen = { version = "0.66.1", features = ["experimental"] }
